+++ CHANGELOG.md
@@ -2,8 +2,8 @@ With a few exceptions, this is a record of mainly just user-facing
 changes—that is, either changes to the actual behavior of the checker, or
 changes to any options/interfaces the checker exposes for developers.
 
-# 17.N.N
-NN XXX 2017
+# 17.11.0
+06 October 2017
   - Allow DPUB roles on more elements (per ARIA in HTML spec updates)
   - Add `--Werror` option to the vnu.jar command-line checker. The option
     causes the checker to exit non-zero if any warnings are encountered